Issue #6439: move powermock tests away from the normal tests
