+++ src/test/java/com/puppycrawl/tools/checkstyle/DefaultLoggerTest.java
@@ -23,9 +23,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 
 import java.io.ByteArrayOutputStream;
 import java.io.OutputStream;
@@ -37,7 +34,6 @@ import org.junit.Test;
 import com.puppycrawl.tools.checkstyle.api.AuditEvent;
 import com.puppycrawl.tools.checkstyle.api.AutomaticBean;
 import com.puppycrawl.tools.checkstyle.api.LocalizedMessage;
-import com.puppycrawl.tools.checkstyle.utils.CommonUtil;
 
 public class DefaultLoggerTest {
 
@@ -74,43 +70,6 @@ public class DefaultLoggerTest {
         dl.addException(new AuditEvent(5000, "myfile"), new IllegalStateException("upsss"));
         dl.auditFinished(new AuditEvent(6000, "myfile"));
         final String output = infoStream.toString();
-        assertTrue("Message should contain exception info, but was " + output,
-                output.contains("java.lang.IllegalStateException: upsss"));
-    }
-
-    @Test
-    public void testNewCtor() throws Exception {
-        final OutputStream infoStream = spy(new ByteArrayOutputStream());
-        final ByteArrayOutputStream errorStream = spy(new ByteArrayOutputStream());
-        final DefaultLogger dl = new DefaultLogger(infoStream,
-                AutomaticBean.OutputStreamOptions.CLOSE, errorStream,
-                AutomaticBean.OutputStreamOptions.CLOSE);
-        dl.auditStarted(null);
-        dl.addException(new AuditEvent(5000, "myfile"), new IllegalStateException("upsss"));
-        dl.auditFinished(new AuditEvent(6000, "myfile"));
-        final String output = errorStream.toString(StandardCharsets.UTF_8.name());
-        final LocalizedMessage addExceptionMessage = new LocalizedMessage(1,
-                Definitions.CHECKSTYLE_BUNDLE, DefaultLogger.ADD_EXCEPTION_MESSAGE,
-                new String[] {"myfile"}, null,
-                getClass(), null);
-        final LocalizedMessage startMessage = new LocalizedMessage(1,
-                Definitions.CHECKSTYLE_BUNDLE, DefaultLogger.AUDIT_STARTED_MESSAGE,
-                CommonUtil.EMPTY_STRING_ARRAY, null,
-                getClass(), null);
-        final LocalizedMessage finishMessage = new LocalizedMessage(1,
-                Definitions.CHECKSTYLE_BUNDLE, DefaultLogger.AUDIT_FINISHED_MESSAGE,
-                CommonUtil.EMPTY_STRING_ARRAY, null,
-                getClass(), null);
-
-        verify(infoStream, times(1)).close();
-        verify(errorStream, times(1)).close();
-        final String infoOutput = infoStream.toString();
-        assertTrue("Message should contain exception info, but was " + infoOutput,
-                infoOutput.contains(startMessage.getMessage()));
-        assertTrue("Message should contain exception info, but was " + infoOutput,
-                infoOutput.contains(finishMessage.getMessage()));
-        assertTrue("Message should contain exception info, but was " + output,
-                output.contains(addExceptionMessage.getMessage()));
         assertTrue("Message should contain exception info, but was " + output,
                 output.contains("java.lang.IllegalStateException: upsss"));
     }
