+++ src/test/java/com/puppycrawl/tools/checkstyle/api/LocalizedMessageTest.java
@@ -23,18 +23,8 @@ import static com.puppycrawl.tools.checkstyle.utils.CommonUtil.EMPTY_OBJECT_ARRA
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyInt;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.powermock.api.mockito.PowerMockito.mock;
-import static org.powermock.api.mockito.PowerMockito.when;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.net.URL;
-import java.net.URLConnection;
-import java.net.URLStreamHandler;
 import java.util.Locale;
 import java.util.Map;
 import java.util.ResourceBundle;
@@ -80,60 +70,6 @@ public class LocalizedMessageTest {
                 Locale.ENGLISH, "java.class",
                 Thread.currentThread().getContextClassLoader(), true);
         assertNull("Bundle should be null when reload is true and URL is null", bundle);
-    }
-
-    @Test
-    public void testBundleReloadUrlNotNull() throws IOException {
-        final ClassLoader classloader = mock(ClassLoader.class);
-        final String resource =
-                "com/puppycrawl/tools/checkstyle/checks/coding/messages_en.properties";
-        final URLConnection mockUrlCon = mock(URLConnection.class);
-        final URLStreamHandler stubUrlHandler = new URLStreamHandler() {
-            @Override
-            protected URLConnection openConnection(URL u) {
-                return mockUrlCon;
-            }
-        };
-        final URL url = new URL("foo", "bar", 99, "/foobar", stubUrlHandler);
-        final InputStream inputStreamMock = mock(InputStream.class);
-        when(classloader.getResource(resource)).thenReturn(url);
-        when(mockUrlCon.getInputStream()).thenReturn(inputStreamMock);
-        when(inputStreamMock.read(any(), anyInt(), anyInt())).thenReturn(-1);
-
-        final LocalizedMessage.Utf8Control control = new LocalizedMessage.Utf8Control();
-        control.newBundle("com.puppycrawl.tools.checkstyle.checks.coding.messages",
-                Locale.ENGLISH, "java.class",
-                classloader, true);
-
-        verify(mockUrlCon, times(1)).setUseCaches(false);
-        verify(inputStreamMock, times(1)).close();
-    }
-
-    @Test
-    public void testBundleReloadUrlNotNullStreamNull() throws IOException {
-        final ClassLoader classloader = mock(ClassLoader.class);
-        final String resource =
-            "com/puppycrawl/tools/checkstyle/checks/coding/messages_en.properties";
-
-        final URL url = getMockUrl(null);
-        when(classloader.getResource(resource)).thenReturn(url);
-
-        final LocalizedMessage.Utf8Control control = new LocalizedMessage.Utf8Control();
-        final ResourceBundle bundle = control.newBundle(
-                "com.puppycrawl.tools.checkstyle.checks.coding.messages",
-                Locale.ENGLISH, "java.class",
-                classloader, true);
-        assertNull("Bundle should be null when stream is null", bundle);
-    }
-
-    private static URL getMockUrl(final URLConnection connection) throws IOException {
-        final URLStreamHandler handler = new URLStreamHandler() {
-            @Override
-            protected URLConnection openConnection(final URL url) {
-                return connection;
-            }
-        };
-        return new URL("http://foo.bar", "foo.bar", 80, "", handler);
     }
 
     @Test
