+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/TranslationCheckTest.java
@@ -23,35 +23,22 @@ import static com.puppycrawl.tools.checkstyle.checks.TranslationCheck.MSG_KEY;
 import static com.puppycrawl.tools.checkstyle.checks.TranslationCheck.MSG_KEY_MISSING_TRANSLATION_FILE;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.endsWith;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.IOException;
 import java.io.Writer;
 import java.lang.reflect.Field;
-import java.lang.reflect.Method;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Set;
-import java.util.SortedSet;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
-import org.junit.runner.RunWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
-import org.mockito.Mockito;
-import org.powermock.modules.junit4.PowerMockRunner;
 
 import com.google.common.collect.ImmutableMap;
 import com.puppycrawl.tools.checkstyle.AbstractXmlTestSupport;
@@ -63,18 +50,13 @@ import com.puppycrawl.tools.checkstyle.api.AutomaticBean;
 import com.puppycrawl.tools.checkstyle.api.Configuration;
 import com.puppycrawl.tools.checkstyle.api.FileText;
 import com.puppycrawl.tools.checkstyle.api.LocalizedMessage;
-import com.puppycrawl.tools.checkstyle.api.MessageDispatcher;
 import com.puppycrawl.tools.checkstyle.internal.utils.XmlUtil;
 import com.puppycrawl.tools.checkstyle.utils.CommonUtil;
 
-@RunWith(PowerMockRunner.class)
 public class TranslationCheckTest extends AbstractXmlTestSupport {
 
     @Rule
     public final TemporaryFolder temporaryFolder = new TemporaryFolder();
-
-    @Captor
-    private ArgumentCaptor<SortedSet<LocalizedMessage>> captor;
 
     @Override
     protected String getPackageLocation() {
@@ -234,55 +216,6 @@ public class TranslationCheckTest extends AbstractXmlTestSupport {
             propertyFiles,
             getPath("app-dev.properties"),
             expected);
-    }
-
-    @Test
-    @SuppressWarnings("unchecked")
-    public void testLogIoExceptionFileNotFound() throws Exception {
-        //I can't put wrong file here. Checkstyle fails before check started.
-        //I saw some usage of file or handling of wrong file in Checker, or somewhere
-        //in checks running part. So I had to do it with reflection to improve coverage.
-        final TranslationCheck check = new TranslationCheck();
-        final DefaultConfiguration checkConfig = createModuleConfig(TranslationCheck.class);
-        final MessageDispatcher dispatcher = mock(MessageDispatcher.class);
-        check.configure(checkConfig);
-        check.setMessageDispatcher(dispatcher);
-
-        final Method loadKeys =
-            check.getClass().getDeclaredMethod("getTranslationKeys", File.class);
-        loadKeys.setAccessible(true);
-        final Set<String> keys = (Set<String>) loadKeys.invoke(check, new File(".no.such.file"));
-        assertTrue("Translation keys should be empty when File is not found", keys.isEmpty());
-
-        Mockito.verify(dispatcher, times(1)).fireErrors(any(String.class), captor.capture());
-        final String actual = captor.getValue().first().getMessage();
-        final LocalizedMessage localizedMessage = new LocalizedMessage(1,
-                Definitions.CHECKSTYLE_BUNDLE, "general.fileNotFound",
-                null, null, getClass(), null);
-        assertEquals("Invalid message", localizedMessage.getMessage(), actual);
-    }
-
-    @Test
-    public void testLogIoException() throws Exception {
-        //I can't put wrong file here. Checkstyle fails before check started.
-        //I saw some usage of file or handling of wrong file in Checker, or somewhere
-        //in checks running part. So I had to do it with reflection to improve coverage.
-        final TranslationCheck check = new TranslationCheck();
-        final DefaultConfiguration checkConfig = createModuleConfig(TranslationCheck.class);
-        final MessageDispatcher dispatcher = mock(MessageDispatcher.class);
-        check.configure(checkConfig);
-        check.setMessageDispatcher(dispatcher);
-
-        final Method logException = check.getClass().getDeclaredMethod("logException",
-                Exception.class,
-                File.class);
-        logException.setAccessible(true);
-        final File file = new File("");
-        logException.invoke(check, new IOException("test exception"), file);
-
-        Mockito.verify(dispatcher, times(1)).fireErrors(any(String.class), captor.capture());
-        final String actual = captor.getValue().first().getMessage();
-        assertThat("Invalid message: " + actual, actual, endsWith("test exception"));
     }
 
     @Test
