+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/javadoc/ClassResolverTest.java
@@ -20,23 +20,14 @@
 package com.puppycrawl.tools.checkstyle.checks.javadoc;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
-import static org.mockito.ArgumentMatchers.any;
 
 import java.util.HashSet;
 import java.util.Set;
 
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.powermock.api.mockito.PowerMockito;
-import org.powermock.core.classloader.annotations.PrepareForTest;
-import org.powermock.modules.junit4.PowerMockRunner;
 
-@RunWith(PowerMockRunner.class)
-@PrepareForTest({ ClassResolver.class, ClassResolverTest.class })
 public class ClassResolverTest {
 
     @Test
@@ -118,54 +109,6 @@ public class ClassResolverTest {
             // exception is expected
             assertEquals("Invalid exception message", "Entry", ex.getMessage());
         }
-    }
-
-    @Test
-    public void testResolveQualifiedNameFails() throws Exception {
-        final Set<String> imports = new HashSet<>();
-        imports.add("java.applet.someClass");
-
-        final ClassResolver classResolver = PowerMockito.spy(new ClassResolver(Thread
-                .currentThread().getContextClassLoader(), "", imports));
-
-        PowerMockito.doThrow(new ClassNotFoundException("expected exception"))
-                .when(classResolver, "safeLoad", any());
-        PowerMockito.doReturn(true).when(classResolver, "isLoadable", any());
-
-        try {
-            classResolver.resolve("someClass", "");
-            fail("IllegalStateException is expected");
-        }
-        catch (IllegalStateException ex) {
-            // exception is expected
-            final String expected = "expected exception";
-            assertTrue("Invalid exception cause, should be: ClassNotFoundException",
-                    ex.getCause() instanceof ClassNotFoundException);
-            assertTrue("Invalid exception message, should end with: " + expected,
-                    ex.getMessage().endsWith(expected));
-        }
-    }
-
-    /**
-     * This test exists to prevent any possible regression and let of
-     * https://github.com/checkstyle/checkstyle/issues/1192 to be persistent
-     * event is not very obvious.
-     *
-     * @throws Exception when smth is not expected
-     */
-    @Test
-    public void testIsLoadableWithNoClassDefFoundError() throws Exception {
-        final Set<String> imports = new HashSet<>();
-        imports.add("java.applet.someClass");
-
-        final ClassResolver classResolver = PowerMockito.spy(new ClassResolver(Thread
-                .currentThread().getContextClassLoader(), "", imports));
-
-        PowerMockito.doThrow(new NoClassDefFoundError("expected exception"))
-                .when(classResolver, "safeLoad", any());
-
-        final boolean result = classResolver.isLoadable("someClass");
-        assertFalse("result should be false", result);
     }
 
 }
