+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/javadoc/JavadocPackageCheckTest.java
@@ -23,8 +23,6 @@ import static com.puppycrawl.tools.checkstyle.checks.javadoc.JavadocPackageCheck
 import static com.puppycrawl.tools.checkstyle.checks.javadoc.JavadocPackageCheck.MSG_PACKAGE_INFO;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.when;
 
 import java.io.File;
 import java.util.Collections;
@@ -129,27 +127,6 @@ public class JavadocPackageCheckTest
             getPath("annotation"
                     + File.separator + "package-info.java"),
             getPath("annotation"
-                    + File.separator + "package-info.java"), expected);
-    }
-
-    /**
-     * Test require readable file with no parent to be used.
-     * Usage of Mockito.spy() is the only way to satisfy these requirements
-     * without the need to create new file in current working directory.
-     *
-     * @throws Exception if error occurs
-     */
-    @Test
-    public void testWithFileWithoutParent() throws Exception {
-        final DefaultConfiguration moduleConfig = createModuleConfig(JavadocPackageCheck.class);
-        final File fileWithoutParent = spy(new File(getPath("noparentfile"
-                    + File.separator + "package-info.java")));
-        when(fileWithoutParent.getParent()).thenReturn(null);
-        when(fileWithoutParent.getParentFile()).thenReturn(null);
-        final String[] expected = CommonUtil.EMPTY_STRING_ARRAY;
-        verify(createChecker(moduleConfig),
-                new File[] {fileWithoutParent},
-                getPath("annotation"
                     + File.separator + "package-info.java"), expected);
     }
